+++ qiskit/synthesis/two_qubits/xx_decompose/polytopes.py
