+++ qiskit/transpiler/passes/synthesis/unitary_synthesis.py
@@ -25,7 +25,7 @@ from qiskit.transpiler.basepasses import TransformationPass
 from qiskit.transpiler.exceptions import TranspilerError
 from qiskit.dagcircuit.dagcircuit import DAGCircuit
 from qiskit.synthesis.one_qubit import one_qubit_decompose
-from qiskit.quantum_info.synthesis.xx_decompose import XXDecomposer, XXEmbodiments
+from qiskit.synthesis.two_qubits.xx_decompose import XXDecomposer, XXEmbodiments
 from qiskit.quantum_info.synthesis.two_qubit_decompose import (
     TwoQubitBasisDecomposer,
     TwoQubitWeylDecomposition,
