+++ test/python/synthesis/xx_decompose/test_polytopes.py
@@ -21,7 +21,7 @@ import ddt
 import numpy as np
 from numpy import pi
 
-from qiskit.quantum_info.synthesis.xx_decompose.polytopes import XXPolytope
+from qiskit.synthesis.two_qubits.xx_decompose.polytopes import XXPolytope
 
 EPSILON = 0.001
